refactor(plan): migrate changed files to path aliases (@/ and @test/)
Converts all deep-relative imports (../../src/, ../../../../src/) in the
files changed on this branch to use path aliases, keeping the
deep-relatives count below the baseline (3319 → 3228, -91 migrated).

Barrel additions required to support the alias migrations:
- src/agents/index.ts: export SessionHandle, TurnResult, RunAsSessionOpts,
  RetryStrategy, RetryPreset, RetryContext, RetryDecision
- src/cli/index.ts: export _planDeps, PlanCommandOptions
- src/config/index.ts: export COMPLEXITY_GUIDE, TEST_STRATEGY_GUIDE,
  SPEC_ANCHOR_RULES, DESCRIPTION_QUALITY_RULES, GROUPING_RULES,
  getAcQualityRules, resolveTestStrategy, VALID_TEST_STRATEGIES
- src/prd/index.ts: export validatePlanOutput
