+++ src/cli/index.ts
@@ -2,7 +2,8 @@
  * CLI Utilities
  */
 
-export { planCommand } from "./plan";
+export { planCommand, _planDeps } from "./plan";
+export type { PlanCommandOptions } from "./plan";
 export { planDecomposeCommand, runReplanLoop } from "./plan-decompose";
 export { acceptCommand, type AcceptOptions } from "./accept";
 export {
