+++ test/unit/cli/plan-debate.test.ts
@@ -15,12 +15,12 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import { mkdir } from "node:fs/promises";
 import { join } from "node:path";
-import { _planDeps, planCommand } from "../../../src/cli/plan";
+import { _planDeps, planCommand } from "@/cli";
 import type { NaxConfig } from "../../../src/config";
-import type { DebateResult } from "../../../src/debate/types";
-import type { PRD } from "../../../src/prd/types";
-import { cleanupTempDir, makeTempDir } from "../../helpers/temp";
-import { makeMockAgentManager, makeMockRuntime, makeNaxConfig } from "../../helpers";
+import type { DebateResult } from "@/debate/types";
+import type { PRD } from "@/prd/types";
+import { cleanupTempDir, makeTempDir } from "@test/helpers";
+import { makeMockAgentManager, makeMockRuntime, makeNaxConfig } from "@test/helpers";
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Helpers
