+++ test/unit/cli/plan-interactive.test.ts
@@ -8,11 +8,11 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import { mkdir, rm } from "node:fs/promises";
 import { join } from "node:path";
-import { _planDeps as _deps, planCommand } from "../../../src/cli/plan";
-import { DEFAULT_CONFIG } from "../../../src/config";
-import type { PRD } from "../../../src/prd/types";
-import { makeMockAgentManager, makeMockRuntime } from "../../helpers";
-import { makeTempDir } from "../../helpers/temp";
+import { _planDeps as _deps, planCommand } from "@/cli";
+import { DEFAULT_CONFIG } from "@/config";
+import type { PRD } from "@/prd/types";
+import { makeMockAgentManager, makeMockRuntime } from "@test/helpers";
+import { makeTempDir } from "@test/helpers";
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Helpers
