+++ test/unit/cli/plan-monorepo.test.ts
@@ -11,10 +11,10 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import { mkdir, rm } from "node:fs/promises";
 import { join } from "node:path";
-import { _planDeps, planCommand } from "../../../src/cli/plan";
-import type { PRD } from "../../../src/prd/types";
-import { makeTempDir } from "../../helpers/temp";
-import { makeMockAgentManager, makeMockRuntime, makeNaxConfig } from "../../helpers";
+import { _planDeps, planCommand } from "@/cli";
+import type { PRD } from "@/prd/types";
+import { makeTempDir } from "@test/helpers";
+import { makeMockAgentManager, makeMockRuntime, makeNaxConfig } from "@test/helpers";
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Helpers
